requirements.txt
aiogram==3.20.0.post0
asyncpg==0.30.0
geopy==2.4.1
Pillow==11.2.1
python-dotenv==1.1.0
Unidecode==1.4.0
